requirements.txt
keyboard==0.13.5
mss==10.0.0
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.0.0
PyDirectInput==1.0.4
pyinstaller==6.11.1
PyWinCtl==0.4.1
PyYAML==6.0.2
typing_extensions==4.12.2
